+++ Questionable/Controller/Steps/Shared/WaitAtEnd.cs
@@ -22,6 +22,7 @@ internal static class WaitAtEnd
         IObjectTable objectTable,
         ICondition condition,
         AutoDutyIpc autoDutyIpc,
+        IAutoHookIpc autoHookIpc,
         BossModIpc bossModIpc,
         RedoUtil redoUtil,
         QuestData questData,
@@ -60,6 +61,7 @@ internal static class WaitAtEnd
 
                 case EInteractionType.Duty when !autoDutyIpc.IsConfiguredToRunContent(step.DutyOptions):
                 case EInteractionType.SinglePlayerDuty when !bossModIpc.IsConfiguredToRunSoloInstance(quest.Id, step.SinglePlayerDutyOptions):
+                case EInteractionType.Fish when !autoHookIpc.IsAvailable():
                     return [new EndAutomation()];
 
                 case EInteractionType.WalkTo: