+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- Bug fix: Mount check fixed -alydev
-- Bug fix for the bug fix: Better logic for using vesper bay ticket
-  - fix bug affecting all UseItem calls, sorry -alydev
+- Feature: AutoHook support -raitono